chore: Organize torin tests
