+++ crates/torin/src/lib.rs
@@ -20,3 +20,5 @@ pub mod prelude {
     pub use crate::torin::*;
     pub use crate::values::prelude::*;
 }
+
+pub mod test_utils;
